backend/requirements.txt

attrdict==2.0.1
cached-property==1.5.2
certifi==2022.12.7
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
coincurve==15.0.1
coloredlogs==15.0.1
configparser==5.0.2
coverage==5.5
cycler==0.10.0
cytoolz==0.11.0
DateTime==4.3
dictionaries==0.0.2
docutils==0.15.2
future==0.18.2
hexbytes==0.2.2
humanfriendly==9.2
humanize==3.11.0
lru-dict==1.1.7
MarkupSafe==2.0.1
mock==4.0.3
mypy-extensions==0.4.3
nltk==3.2.4
numpy==1.21.0
parsimonious==0.8.1
pbkdf2==1.3
persistent==4.7.0
Pillow==9.3.0
plyvel==1.3.0
preprocessing==0.1.13
psycopg2==2.8.5
psycopg2-binary==2.9.1
py-ecc==1.6.0
py-evm==0.3.0a13
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.0.0
pycparser==2.20
pycryptodome==3.10.1
pyethash==0.1.27
pyparsing==2.4.7
pysha3==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
repoze.lru==0.7
requests==2.24.0
rlp==1.2.0
scrypt==0.8.18
semantic-version==2.8.5
simplejson==3.17.2
transaction==3.0.1
urllib3

